fix(suite): unit test setup
