+++ packages/suite/src/actions/settings/__tests__/deviceSettingsActions.test.ts
@@ -1,3 +1,4 @@
+import { prepareSuiteSettingsReducer, suiteSettingsInitialState } from '@suite/settings';
 import { deviceActions } from '@suite-common/device';
 import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { filterThunkActionTypes } from '@suite-common/test-utils';
@@ -5,6 +6,7 @@ import { initialWalletSettingsState } from '@suite-common/wallet-core';
 import TrezorConnect from '@trezor/connect';
 
 import suiteReducer from 'src/reducers/suite/suiteReducer';
+import { extraDependencies } from 'src/support/extraDependencies';
 import { configureStore } from 'src/support/tests/configureStore';
 
 import fixtures, {
@@ -13,10 +15,15 @@ import fixtures, {
 } from '../__fixtures__/deviceSettingsActions';
 
 const DEVICE = mockSuiteDevice({ path: '1', connected: true });
+const suiteSettingsReducer = prepareSuiteSettingsReducer(extraDependencies);
 
 const getInitialState = (state: Partial<DeviceSettingsFixtureState> = {}) => ({
     suite: {
         ...suiteReducer(undefined, { type: '@suite/init' }),
+    },
+    suiteSettings: {
+        ...suiteSettingsInitialState,
+        ...state.suiteSettings,
     },
     device: {
         devices: state.device?.devices ?? [DEVICE],
@@ -37,9 +44,10 @@ const initStore = (state: DeviceSettingsFixtureState) => {
     const store = mockStore(state);
     store.subscribe(() => {
         const action = store.getActions().pop();
-        const { suite, device } = store.getState();
+        const { suite, suiteSettings, device } = store.getState();
         // process action in reducers
         store.getState().suite = suiteReducer(suite, action);
+        store.getState().suiteSettings = suiteSettingsReducer(suiteSettings, action);
         store.getState().device = deviceReducer(device, action);
         // add action back to stack
         store.getActions().push(action);
