+++ packages/suite/src/actions/wallet/__fixtures__/receiveActions.ts
@@ -163,8 +163,8 @@ export default [
     {
         description: 'Show address, device not connected',
         initialState: {
-            suite: {
-                settings: { debug: {} },
+            suiteSettings: {
+                debug: {},
             },
             device: {
                 selectedDevice: mockSuiteDevice({ connected: false }),
