+++ packages/suite/src/hooks/wallet/__fixtures__/useSendForm.ts
@@ -1,6 +1,7 @@
 import { combineReducers, createReducer } from '@reduxjs/toolkit';
 
 import { locksReducer } from '@suite/locks';
+import { suiteSettingsInitialState } from '@suite/settings';
 import { type SuiteSyncDataState, type SuiteSyncState } from '@suite-common/suite-sync';
 import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { testMocks } from '@suite-common/test-utils';
@@ -272,13 +273,13 @@ export const getRootReducer: any = (selectedAccount = BTC_ACCOUNT, fees = DEFAUL
         suite: createReducer(
             {
                 online: true,
-                settings: { debug: {}, theme: { variant: 'light' } },
                 evmSettings: { confirmExplanationModalClosed: {}, explanationBannerClosed: {} },
                 prefillFields: { sendForm: '', transactionHistory: '' },
                 countryCode: null,
             },
             () => ({}),
         ),
+        suiteSettings: createReducer(suiteSettingsInitialState, state => state),
         locks: locksReducer,
         flags: createReducer(
             { stakeEthBannerClosed: false, stakeSolBannerClosed: false },
