+++ suite/settings/src/index.ts
@@ -1,1 +1,2 @@
 export * from './settingsSlice';
+export * from './settingsSelectors';
